Separate Message and Future to different files
